+++ src/main/java/cz/gymtrebon/zaverecky/vjanecek/atlas/repository/RequestImageRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RequestImageRepository extends JpaRepository<RequestImage, Integer> {
+public interface RequestImageRepository extends JpaRepository<RequestImage, Long> {
 }
